Add mock_customer fixture to tests (#1036)
* use mock_customer fixture

* update changelog

* lint

* add fixture to test_to_parquet_with_lti

* lint

* add mock_customer fixture to conftest.py

* call fixture from conftest

* change pd_mock_customer_es to pd_mock_customer

* add username to changelog

* move dd_mock and mock_customer to conftest

* replace pd_mock_customer with mock_customer

* replace pd_mock_customer with dd_mock_customer

* delete dask_mock_customer_es fixture

* lint

* delete mock_customer_es fixture

* replace mock_customer_es with mock_customer

* lint and edit var names

* update changelog future release

* remove space

* add blank line
