+++ featuretools/tests/demo_tests/test_demo_data.py
@@ -2,7 +2,7 @@ import urllib.request
 
 import pytest
 
-from featuretools.demo import load_flight, load_mock_customer, load_retail
+from featuretools.demo import load_flight, load_retail
 from featuretools.synthesis import dfs
 
 
@@ -22,8 +22,8 @@ def test_load_retail_diff():
     assert es_second['order_products'].df.shape[0] == nrows_second
 
 
-def test_mock_customer():
-    es = load_mock_customer(return_entityset=True)
+def test_mock_customer(mock_customer):
+    es = mock_customer
     fm, fl = dfs(entityset=es, target_entity="customers", max_depth=3)
     for feature in fl:
         assert feature.get_name() in fm.columns
