+++ featuretools/tests/entityset_tests/test_serialization.py
@@ -6,7 +6,6 @@ import pandas as pd
 import pytest
 from dask import dataframe as dd
 
-from featuretools.demo import load_mock_customer
 from featuretools.entityset import EntitySet, deserialize, serialize
 from featuretools.variable_types import (
     Categorical,
@@ -214,8 +213,8 @@ def test_to_parquet_interesting_values(pd_es, tmpdir):
     assert pd_es.__eq__(new_es, deep=True)
 
 
-def test_to_parquet_with_lti(tmpdir):
-    es = load_mock_customer(return_entityset=True, random_seed=0)
+def test_to_parquet_with_lti(tmpdir, pd_mock_customer):
+    es = pd_mock_customer
     es.to_parquet(str(tmpdir))
     new_es = deserialize.read_entityset(str(tmpdir))
     assert es.__eq__(new_es, deep=True)
